简单鉴权

+++ controllers/campaign.js
@@ -22,9 +22,9 @@ exports.getAllCampaigns = async (req, res, next) => {
 // 创建新的面试
 exports.createNewCampaign = async (req, res, next) => {
     try {
-        // if (!req.user.groups.some(g => g === 'gdmu/ACM-presidency' || g === 'gdmu/NA-presidency')) {
-        //     throw new AppError('您没有权限创建面试', 403, 'NO_PERMISSION');
-        // }
+        if (!req.user.groups.some(g => g === 'gdmu/ACM-presidency' || g === 'gdmu/NA-presidency' || g === 'gdmu/Nekorify-admin')) {
+            throw new AppError('您没有权限创建面试', 403, 'NO_PERMISSION');
+        }
         const data = req.body;
         const newCampaign = await campaignService.createNewCampaign(data);
         return res.success(newCampaign, '面试创建成功', 'CAMPAIGN_CREATED');
@@ -36,9 +36,9 @@ exports.createNewCampaign = async (req, res, next) => {
 // 更新面试信息
 exports.updateCampaign = async (req, res, next) => {
     try {
-        // if (!req.user.groups.some(g => g === 'gdmu/ACM-presidency' || g === 'gdmu/NA-presidency')) {
-        //     throw new AppError('您没有权限更新面试', 403, 'NO_PERMISSION');
-        // }
+        if (!req.user.groups.some(g => g === 'gdmu/ACM-presidency' || g === 'gdmu/NA-presidency' || g === 'gdmu/Nekorify-admin')) {
+            throw new AppError('您没有权限更新面试', 403, 'NO_PERMISSION');
+        }
         const campaignId = req.params.id;
         const data = req.body;
         console.log('123456789', data);
@@ -53,9 +53,9 @@ exports.updateCampaign = async (req, res, next) => {
 // 删除面试
 exports.deleteCampaign = async (req, res, next) => {
     try {
-        // if (!req.user.groups.some(g => g === 'gdmu/ACM-presidency' || g === 'gdmu/NA-presidency')) {
-        //     throw new AppError('您没有权限删除面试', 403, 'NO_PERMISSION');
-        // }
+        if (!req.user.groups.some(g => g === 'gdmu/ACM-presidency' || g === 'gdmu/NA-presidency' || g === 'gdmu/Nekorify-admin')) {
+            throw new AppError('您没有权限删除面试', 403, 'NO_PERMISSION');
+        }
         const campaignId = req.params.id;
         await campaignService.deleteCampaign(campaignId);
         return res.success(null, '面试删除成功', 'CAMPAIGN_DELETED');
